+++ Backend/Game/src/main/java/com/ygmg/game/api/request/AreaModifyPutReq.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AreaModifyPutReq {
-    int areaId;
+    Long areaId;
     double areaSize;
 }
